Work towards: rdar://107140336 (FoundationPreview: Batch move string API)
Move case mapping API to FoundationEssentials and FoundationInternationalization:
- Essentials: Unlocalized caplitalized
- Internationalization: Localized caplitalized, lowercased, uppercased

Note that in the tests, we test the internal implementation instead of the public function. This is because these functions are declared as public function of
`extension StringProtocol`, so there are two declarations on Darwin: once in FoundationPreview and once in system SDK. There is currently no way to specify which one is intended when building the package on Darwin.