rdar://103959505 (Formatted functions ignore user's date preferences)
If locale's preferences specifies a custom date format for the given date style, use what's given as-is instead of generating a date pattern from the date style. Realistically this only applies to `Locale.current` since that's the only instance that would have embedded preferences.

Encode and decode `_dateStyle` so it survives archiving. We don't resolve the literal format until `.format()` is called. So if the decoded style is modified in a way that the locale override no longer applies, the encoded `_dateStyle` would not be used.