Update `Date.description` tests for Foundation Framework
We use ICU for date calculation for Foundation Framework, and `gmtime` otherwise. It appears that the result starts to diverge at some time in Gregorian year 1582, which happens to be the year when Gregorian calendar went into effect. That in turn results in disparate representations of `Date` before year 1582.

It's not clear what `tm` and `strtime` handles before-Gregorian dates since it's entirely in the implementation details. This PR simply update test expectations accordingly.