Implement Date.description in pure Swift
# Motivation

The current implementation of `Date.description` is depending on the what platform we are running on. On Darwin it used the CFFoundation backed implementation and on Linux/Windows it used libc APIs to do the date calculations. This caused build issues on Windows and the libc APIs are not 32bit clean.

# Changes

This PR changes the implementation to use the `FoundationEssentials` gregorian calendar and ISO 8601 formatter.

# Result

Swift only implementation that works cross platform for `Date.description` which also should perform better than any libc API calls.