Implement getting calendar symbols (#181)
We should include the calendar identifier via the "ca" keyword when getting calendar symbols from ICU, such as "en_US@ca=hebrew". It was mistakenly dropped when we rewrote Calendar, so now we always return the symbols of the default calendar of the locale.

We can fix this by updating the existing implementation with the correct identifier, but we can also just use `ICUDateFormatter`, bypassing `CFDateFormatter` altogether. We chose the latter approach here.

Fixes rdar://110585721