Introduce Locale.identifierCapturingPreferences to capture preferences in a single identifier (#180)
* Part 1: Clean up various usages of force-12-hour and force-24-hour

Consolidate `force24Hour` and `force12Hour` to `forceHourCycle`. Use this as the universal override option.

Work towards resolving rdar://110211953

* Part 2: Introduce `identifierCapturingPreferences`

Add `Locale.identifierCapturingPreferences` to return a string identifier with preferences encoded in. This allows us to pass this identifier to ICU functions to request overrides rather than specifying options in individual API. This PR also adopts this identifier for `ICUNumberFormatter`.

* Add a comment to clarify the variable only includes a subset of preferences