rdar://111859067 (LocalizedStringResource doesn’t work when creating a new String catalog manually) (#207)
When Xcode runs an app in a specified locale, it adds an -AppleLanguages command
line argument that overrides the app language in NSArgumentDomain. These command
line arguments are parsed in `-[NSUserDefaults init]` and are thus guaranteed to
be available on subsequent reads through NSUserDefaults. LocalizedStringResource
however is localized based on `Locale.current` which does not use NSUserDefaults
and may bypass the initialization of preferences in NSArgumentDomain. Thus, it's
possible that an app launched through Xcode can have parts of the UI rendered in
the system language instead of the language specified in the scheme.

To fix this, initialize NSArgumentDomain preferences in CFPreferences instead of
in NSUserDefaults. To verify this change in a unit test, we spawn an instance of
aux_runner and specify a -AppleLanguages argument. In the aux client, don't read
NSUserDefaults but assert that Locale.current has been overriden.

With this fixed, remove a "temporary solution" that searched for -AppleLanguages
in `CFBundleCopyUserLanguages()`. This was introduced in e2d2927cbe and featured
the following comment:

   // This is a temporary solution, until the argument domain is moved down into
   CFPreferences

This comment seems to have been deleted some time between 2001 and 2023, but the
temporary solution remained. We no longer need it. While at it, we can also move
the registration for NSInterfaceStyle="macintosh" down to CFPreferences.