(113054230) Remove a Locale test and update imports in performance tests
Remove test_xcode_override_french() from Locale's test. This was added to test a
behaviour in a part of the Foundation framework that's not open-source and isn't
needed in FoundationEssentials.

Also, update the imports for performance tests. They were importing TestSupport,
which was then importing Foundation as `@testable`. In performance tests, we run
against a release build of the framework, so that import couldn't succeed. Since
performance tests are currently not run in FoundationEssentials (see issue #254)
for now we simply import Foundation and XCTest in lieu of TestSupport.

Co-authored-by: Tony Parker <anthony.parker@apple.com>