Renamed the decimalDigitsAndMagnitudePerWord helper to `check` as well (from `test`), moved it up a level, and changed its `magnitude` parameter to be `UInt`.
This way it can be used from additional test cases in future.

The parameter type change is helpful for any future tests that want to use this with types that might not be initialisable from the necessary integer literal (e.g. an arbitrary-precision integer which doesn't support ExpressibleByIntegerLiteral from StaticBigInt and is thus limited to Int, yet its word magnitude will still be UInt as required by the BinaryInteger protocol and its word magnitude value will not fit into an Int).