Switched to using UnsafeMutableBufferPointer.initializeElement(at:to) rather than simple assignment via subscripting, as the latter mysteriously crashes once the buffer gets large enough.
I wasn't able to precisely root-cause the crash, but it does seem like it's something to do with ContiguousArray's assignment operator implementation assuming `self` is valid, which of course it isn't (yet) in this case.  I didn't explore how `initializeElement(at:to)` does the (figuratively) same assignment without running afoul of this, just observed that it evidently does.