Revised the explanatory comment on why `description` cannot be used instead, despite having the same output for many types of BinaryIntegers.
While the prior comment was correct that the output of `description` can change at any time, which is also a problem, it's not the root problem.  `description`'s output is fundamentally not a defined format and there's no guarantee (nor requirement) that any particular BinaryInteger-conforming type use any particular format.  It doesn't matter the the built-in types happen to use the desired format, this code needs to work for _all_ BinaryIntegers.