Changed the "fast path should handle the zero case" assert to a precondition, at @iCharlesHu's request (https://github.com/apple/swift-foundation/pull/262#discussion_r1347867409).
It should never trip as long as `UInt(exactly:)` and the implementation of the concrete type's words property aren't buggy.  If either is, then it's likely that the result of this method will be wrong anyway.  So the benefit of the check is probably just a faster and clearer failure.

But the downside of the precondition check seems quite small - slight increase in code size and a [hopefully very small] runtime cost.