Added tests using the Numberick 3rd party package (if it's available), to cover fixed-width 128-bit and 256-bit integers.
These are conditionalised on the inclusion of Numberick in Package.swift, which is not the default.  Like the existing BigInt tests.

This proved useful as it did reveal a bug - one I'd actually discovered previously but apparently the fix got lost during Git patch shuffling or something - in `magnitudeBitWidth`, whereby it was forgetting to check if the current word is actually non-zero before treating it as relevant (meaning not just zero padding).  This wasn't caught by the BigInt tests because those BigInt / BigUInt are flexible-width and only use as many words as strictly necessary to contain the value.