[Gregorian Calendar] WIP: Implement `ordinality(of:in:for:)` (#336)
* [Gregorian Calendar] Drive-by fix for quarter handling in `dateComponents(from:)`

The current handling for quarter is wrong for leap years. Fix this by reversing the leap year check.

* [Gregorian Calendar] WIP: Implement `ordinality(of:in:for:)`

The implementation follows that in Calendar_ICU. But instead of mutating the stored ICU calendar, we pass around a `Date` and its `DateComponents` to for keeping track of the calculation state.

Unfinished task: There are a few places where we try to find the next date matching a given date component repeatedly in while loops. We need to add failsafe to prevent infinite loop.

* Address compiler warning

* Fix `start(of:at:)`

Similar to #335, it's necessary to only specify relevant components when converting `Date` to `DateComponents` with `dateComponents(:)`. Address the same issue here.