[Gregorian Calendar] Implement `dateComponents(_:from:to:)` (#378)
* [Gregorian Calendar] Implement `dateComponents(_:from:to:)`

Implementation follows that of Calendar_ICU and ICU's Calendar::fieldDifference.

Like many calendar algorithms, we get the difference by iterative search: We start advancing from `start` until we reach or pass `end`, striding using the queried component. If we land on `end` exactly, return the striding count as-is. If we pass `end`, do a binary search to find the optimal stride.

* Support date of year