Incorrect result for secondsFromGMT for remote dates
We're capping the time zone offset calculation to some arbitrary dates, so the results outside of the date range are incorrect.

The range was added originally as a workaround for poor ICU performance in `ucal_get` for 6238412, but since then ICU has added new API, which we also adopted, so it shouldn't been needed now.

Resolves 50554668