CalendarGregorian search functions may loop indefinitely when reaching the capped date
We are capping the input `Date` to the internal upper bound in various internal functions. This caused the incremental search functions to loop indefinitely because the search results are always the same since the effective input never changes.

Remove the capping from internal functions entirely. The capping was added to improving performance when calling into ICU. We should instead move the capping up to Calendar.swift, the call site to reach consistent behavior, but not inside the implementation.

Also fix miscellaneous other problems found in testing

Resolves 122946218