Unexpected behavior when getting a `Date` after setting `day` on a date components
This is due to ICU's handling of ambiguous/conflicting components. If both "day of year" and "day of month" fields are set via `ucal_set`, the newer one takes priority. Therefore it breaks existing behavior when we set "day of year" after having set "day of month".

Fix this by moving "day of year" before "day of month" to maintain compatibility.

Resolves 123703256