Return early if the component value being added is zero (#480)
When adding month or year to a `Date`, we convert a `Date` to a `DateComponent`, modify the component, and convert the modified `DateComponent` back to `Date`. However for dates falling into the "repeated" time frame on DST end date, we do have have the information of which time the `DateComponent` represents when converting it back to `Date`.

This has been the design, and we have always assumed the second occurrence of the same time. The only implementation difference is that in the ICU implementation, we return early when the value is 0 and do not proceed onto adding. We should do that for GregorianCalendar too.

Resolves 124231654