[Gregorian Calendar] Correctly handle last week of year and first week of year when calculating a `Date` from `DateComponents`
When deciding the component values when calculating `Date` given a `DateComponent`, we incorrectly rewind to the previous year if `weekOfYear == 1` regardless of the passed-in `month`. This adjustment is incorrect and unncessary when `month == 1` -- being in the first month already presumes the start of the year, so doing so results in over-adjustment.

Fix this by skipping this adjustment if we're already in the first month.

Resolves 124661497