_CShims: repair the Windows build
This was obscuring the current foundation blocker for Windows: ICU. The
last update to ICU broke the Windows build and there does not seem to be
a clear repair for that. See apple/swift-foundation-icu#15 for more
details on the ICU failure. In the meantime, repair the `environ`
handling for Windows. `unistd.h` is the Unix standard header which must
be available when `__unix__` is defined. Add an explicit check for that
to avoid including the header on other platforms.