Performance improvements for `Calendar.RecurrenceRule`
This commit improves runtime of `Calendar.RecurrenceRule.occurrences(of:)` by

The original implementation of `Calendar.RecurrenceRule` computed recurrences of
dates using the Calendar APIs for matching date components. Although this worked
well, multiple sequences for matching date components for cases
where a single sequence would have sufficed. This is by design: since each field
of the recurrence rule can have multiple values (e.g. an event repeating Tuesdays and Fridays), we can not represent these in a single DateComponents.

For some cases, this results in up to 4x runtime penalty for finding recurrences as opposed to using Calendar.dates(byMatching:). In the benchmark below computing the next thousand Thanksgivings, we have:

one base sequence to find yearly recurrences of a date (e.g. June 10 2024, June 10 2025)
for each yearly recurrence, we use Calendar.nextDate(after:, matching:) to find a date in the month of November. (e.g. November 10 2024, November 10 2025)
for each date in November, we use component matching to find the fourth Thursday.
This results in a lot of redundant calculations. Since we also own enumerating by date components, we can use some of the internals to more efficiently calculate recurrences of dates, as opposed to using only public API.