Add components in the order of granularity in `date(byAdding:to:)` (#482)
* Add components in the order of granularity in `date(byAdding:to:)`

We compute the differences in Calendar component values from the largest to the smallest in `dateComponents(_:from:to:)`, but we are not following the same order in `date(byAdding:to:)`. Change the order of the latter so we add the largest components first.

The results are the same for most cases regardless of the order. The only exception is when the addition moves the date across DST transition. The reason for this is that the implementation aims to maintain the clock time when adding units that are larger than `day` to a date, so that the time in the day remains unchanged even after time zone offset changes. However, we cannot hold this promise if the result lands in the "skipped hour" on the DST start date as that time does not actually exist. To adjust for this case, we adjust the time of the day to the correct timezone.

For example, the DST start of year 2024 in the Los Angeles time zone is 2024-03-10, where 02:00:00-0800 becomes 03:00:00-0700.

Here's what happens when we add 1 day first, then add 1 week, to 2024-03-09T02:34:36-0800, the day before DST start:

We start by adding 1 day:
2024-03-09T02:34:36-0800 + 1 day --> 2024-03-10T02:34:36-0800

This is not a valid date in this time zone, so we adjust it to -0700:
2024-03-10T02:34:36-0800 --> 2024-03-10T03:34:36-0700

Then we add a week, which is 7 days:
2024-03-10T03:34:36-0700 + 7 days -> 2024-03-17T03:34:36-0700

Notice that the local time is different in the result from that of the original, which changes from 02:34 to 03:34.

If we start by adding 1 week:
2024-03-09T02:34:36-0800 + 7 days -> 2024-03-16T02:34:36-0700

Then add a day:
2024-03-16T02:34:36-0700 + 1 day -> 2024-03-17T02:34:36-0700

The local time is the same as that of the original.

We go for the latter to allow round-tripping a date to the date by adding their differences back.

Resolves 124414807

* Update test and remove unreasonable test cases

Currently there are tests to verify the behavior of adding `DateComponents` that sets both `weekOfYear` and `weekOfMonth`. This configuration is ambiguous, and even more so when `wrapping` is `true` since the results vary depending on which component is added first.

Replace tests asserting the behavior of this unreasonable `DateComponents` with those that only uses one kind of `week` component.

* Review feedback

* Address review feedback

* Review feedback: add back a test case with updated expectation and change the test format