FoundationEssentials: avoid precondition failure on Windows
Repair the error reporting on Windows in the case of a failure in
`SHGetKnownFolder` - the returned value is a `HRESULT`, which is the
extended error code unlike the Win32 APIs which return a `BOOL`.

The "Recyle Bin" is a virtual folder and does not have an associated
path with the folder that we can query. Undo the implementation here for
the time being to avoid the assertion failure in the runtime.