FoundationEssentials: avoid a sharing violation on Windows (#615)
When trying to perform a move, we would encounter a sharing violation
due to the file descriptor being open preventing the file from being
moved. Perform an early close of the file descriptor before performing
the move. This adds complexity to the defer, however, avoiding the defer
complicates the code flow analysis for error propagation.