Some tests about IntegerParseStrategy's big int behavior w.r.t. Double.
IntegerParseStrategy currently tries to parse a Double when it fails to parse an Int64, which results in some odd behavior for magnitudes greater than 2^53. The oddities stem the from fact that the parseAsDouble(_:) method rounds its input String to the nearest Double. This commit adds some tests that fail when this happens.