FoundationEssentials: correct callbacks for `moveItem(atPath:toPath:)` (#661)
In the case that the destination exists, we should check for continue
after error. If the source does not exist, simply do nothing.

Furthermore, adjust the test to ensure that we are using the path
spelling rather than the FSR.

Take the opportunity to correct an error in the tests where we would not
check the correct captured item set.