Added JSON decoding tests asserting that integers are not rounded to the nearest Double.
These tests fail because the most negative signed fixed-width integer is an exact Double, which means that values beyond T.min but closer to T.min than the next Double get clamped to T.min in the JSONDecoder's fallback Double path. Note that unsigned integers don't have this problem because the nearest Double is out of bounds.